+++ bin/run-metrics-util-tests.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+
+set -eo pipefail
+
+cd `dirname $BASH_SOURCE`/../packages/metrics-util
+
+npm run test:debug
+