+++ packages/google-cloud-run/images/instana-google-cloud-run/setup.sh
@@ -1,0 +1,3 @@
+cd /instana
+npm rebuild
+