+++ packages/google-cloud-run/images/test-images/build-and-push.sh
@@ -1,0 +1,32 @@
+#!/usr/bin/env bash
+
+set -eo pipefail
+
+cd `dirname $BASH_SOURCE`
+
+source utils
+
+normalizeArgs $1 $2 $3
+
+if [[ ! -f .env ]]; then
+  echo .env file is missing
+  exit 1
+fi
+source .env
+
+if [[ -z "${image_tag_prefix-}" ]]; then
+  echo Please set image_tag_prefix in .env.
+  exit 1
+fi
+if [[ -z "${container_name_prefix-}" ]]; then
+  echo Please set container_name_prefix in .env.
+  exit 1
+fi
+
+./build.sh $INSTANA_LAYER_MODE $NODEJS_VERSION $LINUX_DISTRIBUTION
+
+setImageTag $image_tag_prefix $NODEJS_VERSION $LINUX_DISTRIBUTION $INSTANA_LAYER_MODE
+
+echo "Pushing image $image_tag to $gcr_repository"
+docker push $gcr_repository/$image_tag
+