+++ packages/metrics-util/src/index.js
@@ -1,0 +1,9 @@
+'use strict';
+
+exports = module.exports = {
+  DataProcessor: require('./DataProcessor'),
+  DataSource: require('./DataSource'),
+  HttpDataSource: require('./HttpDataSource'),
+  nodejs: require('./nodejs'),
+  process: require('./process')
+};