+++ packages/metrics-util/src/nodejs/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+exports = module.exports = {
+  coreAndShared: require('./coreAndShared'),
+  NodeJsProcessor: require('./NodeJsProcessor')
+};