+++ packages/metrics-util/src/process/index.js
@@ -1,0 +1,5 @@
+'use strict';
+
+exports = module.exports = {
+  ProcessProcessor: require('./ProcessProcessor')
+};